Updated hacker news example to new SFA format and added a README
