+++ Cargo.toml
@@ -20,9 +20,7 @@ members = [
   "examples/counters",
   "examples/counters-stable",
   "examples/fetch",
-  "examples/hackernews/hackernews-app",
-  "examples/hackernews/hackernews-client",
-  "examples/hackernews/hackernews-server",
+  "examples/hackernews",
   "examples/parent-child",
   "examples/router",
   "examples/todomvc",
