--- examples/hackernews/hackernews-app/src/main.rs
@@ -1,11 +1,0 @@
-use hackernews_app::*;
-use leptos::*;
-
-pub fn main() {
-    console_error_panic_hook::set_once();
-    _ = console_log::init_with_level(log::Level::Debug);
-    console_error_panic_hook::set_once();
-    mount_to_body(|cx| {
-        view! { cx, <App/> }
-    });
-}
