--- examples/hackernews/hackernews-client/build.sh
@@ -1,1 +1,0 @@
-wasm-pack build --target=web --release
