--- examples/hackernews/hackernews-server/.gitignore
@@ -1,2 +1,0 @@
-key.pem
-cert.pem
