--- examples/hackernews/hackernews-server/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "hackernews-server"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-actix-files = "0.6"
-actix-web = { version = "4", features = ["openssl", "macros"] }
-futures = "0.3"
-leptos = { path = "../../../leptos", default-features = false, features = ["ssr", "serde"] }
-leptos_router = { path = "../../../router", default-features = false, features = ["ssr"] }
-leptos_meta = { path = "../../../meta", default-features = false, features = ["ssr"] }
-log = "0.4"
-hackernews-app = { path = "../hackernews-app", default-features = false, features = ["ssr"] }
-openssl = { version = "0.10", features = ["v110"] }
-simple_logger = "2"
-serde_json = "1.0.85"
