+++ examples/hackernews/src/api.rs
