+++ examples/hackernews/src/routes.rs
@@ -1,0 +1,4 @@
+pub mod nav;
+pub mod stories;
+pub mod story;
+pub mod users;
