+++ examples/hackernews/src/routes/nav.rs
