+++ examples/hackernews/src/routes/stories.rs
