+++ examples/hackernews/src/routes/story.rs
