+++ examples/hackernews/src/routes/users.rs
